adicionado funcionalidade de editar msgs em servidores, tentei mais falhei miseravelmente em fazer um sistema de notificação, to mt cansado pra colocar tudo aqui

+++ app/components/main/main/messagesBox.tsx
@@ -35,7 +35,8 @@ export default function messagesBox() {
                     timestamp: Timestamp.now(),
                     serverId: serverData.id,
                     chatId: selectedChannel.id,
-                    content: message
+                    content: message,
+                    createdAt:Timestamp.now()
                 });
 
 
@@ -52,7 +53,7 @@ export default function messagesBox() {
                 collection(db, 'mensagens'),
                 where('serverId', '==', serverId),
                 where('chatId', '==', chatId),
-                orderBy('timestamp')
+                orderBy('createdAt')
             );
     
             const unsubscribe = onSnapshot(q, (querySnapshot) => {
@@ -68,38 +69,45 @@ export default function messagesBox() {
     };
 
 
-    const deleteMessage = async (chatId: string | null, messageId: string) => {
+    const deleteMessage = async (chatId: string | null, messageId: string, messageUserId: string) => {
         if (!chatId || !messageId) {
             console.error('chatId e messageId são necessários.');
             return;
         }
-
+    
+        // Verifica se o usuário atual é o proprietário da mensagem
+        if (messageUserId !== userData?.uid) {
+            console.error('Você não tem permissão para deletar esta mensagem.');
+            return;
+        }
+    
         try {
             const q = query(
                 collection(db, 'mensagens'),
                 where('chatId', '==', chatId),
                 where('messageId', '==', messageId)
             );
-
+    
             const querySnapshot = await getDocs(q);
-
+    
             if (querySnapshot.empty) {
                 console.error('Nenhuma mensagem encontrada com o messageId fornecido.');
                 return;
             }
-
+    
             querySnapshot.forEach(async (docSnapshot) => {
                 const docRef = doc(db, 'mensagens', docSnapshot.id);
                 await deleteDoc(docRef);
                 console.log(`Mensagem com ID ${docSnapshot.id} deletada com sucesso.`);
-
-                // Re-fetch the messages after deletion
+    
+                // Atualiza as mensagens após a exclusão
                 await fetchMessages(serverData.id, selectedChannel.id);
             });
         } catch (error) {
             console.error('Erro ao deletar a mensagem: ', error);
         }
     };
+    
 
     // Use o useEffect para buscar mensagens quando serverData e selectedChannel mudam
     useEffect(() => {
@@ -114,7 +122,7 @@ export default function messagesBox() {
 
     async function getMessageRefById(chatId: string, id: unknown) {
         const messagesRef = collection(db, 'mensagens');
-        const q = query(messagesRef, where('id', '==', id));
+        const q = query(messagesRef, where('messageId', '==', id));
         const querySnapshot = await getDocs(q);
 
         if (!querySnapshot.empty) {
@@ -123,11 +131,11 @@ export default function messagesBox() {
         throw new Error('Mensagem não encontrada.');
     }
 
-    async function editMessage(chatId: never, id: null, newMessage: string) {
+    async function editMessage(chatId: never, editMessageId: null, newMessage: string) {
         try {
-            const messageRef = await getMessageRefById(chatId, id);
+            const messageRef = await getMessageRefById(chatId, editMessageId);
             await updateDoc(messageRef, {
-                message: newMessage,
+                content: newMessage,
                 timestamp: serverTimestamp(),
                 isEdited: true // Marcar a mensagem como editada
 
@@ -146,11 +154,11 @@ export default function messagesBox() {
 
     function handleAction(chatId: { key: string; preventDefault: () => void; }, action: string) {
         if (chatId && action === 'post') {
-            sendMessage(chatId, userData.uid, frienduserData.uid, message);
+            sendMessage(chatId);
             setMessage('');
         } else if (chatId && action === 'edit') {
             editMessage(chatId, editMessageId, messageEdited);
-            setIsEditingMessage('');
+            setIsEditingMessage(null);
         } else {
             console.log('erro');
         }
@@ -179,96 +187,123 @@ export default function messagesBox() {
         },
     };
 
+    const formatTimestamp = (timestamp: { seconds: number; }) => {
+        if (!timestamp) return '';
+        const date = new Date(timestamp.seconds * 1000);
+        const now = new Date();
+
+        const dayDiff = Math.floor((now - date) / (1000 * 60 * 60 * 24));
+
+        if (dayDiff === 0) {
+            return date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
+        } else if (dayDiff === 1) {
+            return `Ontem, ${date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })}`;
+        } else {
+            return date.toLocaleDateString([], { day: '2-digit', month: '2-digit', year: 'numeric' }) +
+                ` ${date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })}`;
+        }
+    };
     return (
-        <Box
-            height={["calc(100vh - 50px)"]}
-            display="flex"
-            flexDirection="column"
-            overflow="hidden"
-            width="100%"
-        >
-            <Box
-                css={scrollStyle}
-                flex="1"
-                display="flex"
-                flexDirection="column"
-                justifyContent="flex-end"
-
-                overflowY="auto"
-                width="100%"
-                p="4"
-            >
+        (
+            <Box height={["calc(100vh - 50px)"]} display="flex" flexDirection="column" overflow="hidden" width="100%">
+                <Box css={scrollStyle} flex="1" display="flex" flexDirection="column" justifyContent="flex-end" overflowY="auto" width="100%" p="4">
                 <Stack spacing="3">
-
-                    <Box display={'flex'} justifyContent={'center'} alignItems={'center'} width={'80px'} bg='#383a40' height={'80px'} borderRadius={'80px'}>
-                        <FaHashtag fontSize={'45px'} color='white' />
-                    </Box>
-                    <Text fontSize={'25px'} color={'white'} fontWeight={'800'}>Bem vindo(a) a {selectedChannel?.name}</Text>
-                    <Text fontSize={'15px'} mt={'-5px'} color={'#96989D'}>Esse é o começo do canal {selectedChannel?.name}</Text>
-
-
-                    {getMessages && getMessages.length > 0 && getMessages.map((e, index) => (
-                        <Box w={'100%'} height={'auto'} minH={'55px'} display={'flex'} key={index} onMouseEnter={() => setHoveredIndex(index)}
-                            onMouseLeave={() => setHoveredIndex(null)} _hover={{ 'bg': '#2b2d31' }}
-                            transition={'0.1s all'}
+                    {getMessages.map((msg, index) => (
+                        <Box
+                            w={'100%'}
+                            height={'auto'}
+                            minH={'55px'}
+                            display={'flex'}
+                            key={index}
+                            onMouseEnter={() => setHoveredIndex(index)}
+                            onMouseLeave={() => setHoveredIndex(null)}
+                            _hover={{ 'bg': '#2b2d31' }}
                             alignItems={'center'}
+                            transition={'0.1s all'}
+                            bg={isEditingMessage === index ? '#2b2d31' : 'transparent'}
                         >
                             <Box display={'flex'} justifyContent={'center'} alignItems={'center'} width={'70px'} height={'55px'}>
-                                <Box bg={e?.bgIconColor || '#383a40'} height={'40px'} width={'40px'} borderRadius={'40px'} display={'flex'} justifyContent={'center'} alignItems={'center'}>
+                                <Box
+                                    bg={msg.bgIconColor}
+                                    height={'40px'}
+                                    width={'40px'}
+                                    borderRadius={'40px'}
+                                    display={'flex'}
+                                    justifyContent={'center'}
+                                    alignItems={'center'}
+                                    mb={'3px'}
+                                >
                                     <FaDiscord fontSize={'25px'} color='white' />
                                 </Box>
                             </Box>
-                            <Box flex={'1'}>
-                                <Box width={'100%'} display={'flex'} flexDir={'column'} height={'100%'}>
-                                    <Text color={'white'} as={'span'}>{e?.username}</Text>
-                                    <Text mb={'-2px'} fontSize={'16px'} color={'#96989D'}>{e?.content}</Text>
+                            <Box flex={'1'} height={'100%'}>
+                                <Box width={'100%'} display={'flex'} flexDir={'column'} justifyItems={'center'} height={'100%'}>
+                                    <Text color={'white'} as={'span'}>
+                                        {msg.username}
+                                        <Text ml={'10px'} as={'span'} fontSize={'14px'} color={'#96989D'}>
+                                            {formatTimestamp(msg.timestamp)}
+                                        </Text>
+                                    </Text>
+
+                                    {isEditingMessage === index ? (
+                                        <Box height={'60px'}>
+                                            <Input padding={'0px'} _focusVisible={{ 'border': 'none' }} outline={'none'} color={'white'} bg='#2b2d31' border={'none'} onChange={(e) => setMessageEdited(e.target.value)} value={messageEdited} onKeyDown={(event) => handleKeyDown(msg.chatId, event, 'edit')} />
+                                            <Stack flexDir={'row'}>
+                                                <Text cursor={'pointer'} fontSize={'11px'} color={'white'} onClick={() => [setHoveredIndex(null), console.log(hoveredIndex), setIsEditingMessage(false)]}>Cancelar</Text>
+                                                <Text cursor={'pointer'} fontSize={'11px'} onClick={(event) => handleAction(msg.chatId, 'edit')} color={'#00AFF4'}>Enviar</Text>
+                                            </Stack>
+
+
+                                        </Box>
+                                    ) : (
+                                        <>
+                                            <Text mb={'-2px'} fontSize={'16px'} color={'#96989D'}>
+                                                {msg.content}
+                                                {msg.isEdited && <Text as="span" fontSize={'11px'} color={'#00AFF4'}> (editado)</Text>}
+
+                                            </Text>
+                                        </>
+                                    )}
+
                                 </Box>
                             </Box>
-                            {hoveredIndex === index && !isEditingMessage && (
+                            {hoveredIndex === index && !isEditingMessage  && msg.userId === userData?.uid && (
                                 <>
                                     <MdDelete
                                         cursor={'pointer'}
                                         color='#96989D'
-                                        onClick={() => deleteMessage(e.chatId, e.messageId)}
+                                        onClick={() => deleteMessage(msg.chatId, msg.messageId, msg.userId)}
                                         fontSize={'20px'}
                                         style={{ 'marginRight': '10px' }}
                                     />
                                     <MdEdit
                                         style={{ 'marginRight': '10px' }}
-                                        onClick={() => handleEditMessage(e.messageId, e.content, index)}
+                                        onClick={() => handleEditMessage(msg.messageId, msg.content, index)}
                                         cursor={'pointer'}
                                         color='#96989D'
                                         fontSize={'20px'}
                                     />
                                 </>
                             )}
+
                         </Box>
                     ))}
-
-
-
-
-                    {/* Adicione mais mensagens aqui */}
-
                 </Stack>
+                </Box>
+                <Box p="4" pt={'0px'}>
+                    <Input
+                        bg="#383a40"
+                        placeholder="Digite sua mensagem..."
+                        size="lg"
+                        border="none"
+                        color="white"
+                        onChange={((e) => setMessage(e.target.value))}
+                        value={message}
+                        onKeyDown={sendMessage}
+                        required
+                    />
+                </Box>
             </Box>
-            <Box
-                p="4"
-                pt={'0px'}
+        ))
+    }
 
-            >
-                <Input
-                    bg="#383a40"
-                    placeholder="Digite sua mensagem..."
-                    size="lg"
-                    border="none"
-                    color="white" // Para que o texto seja visível em fundo escuro
-                    onChange={((e) => setMessage(e.target.value))}
-                    value={message}
-                    onKeyDown={sendMessage}
-                    required
-                />
-            </Box>
-        </Box>
-    )
-}
